// src/scraper/types/database/entities.ts
export enum Entities {
  Country = 'countries',
  Stadium = 'stadiums',
  Teams = 'teams',
  Players = 'players',
  Matches = 'matches',
  Positions = 'positions',
  PlayerPositions = 'playerPositions',
  MatchGoals = 'matchGoals',
  MatchAssits = 'matchAssists',
  MatchCards = 'matchCards',
  PlayerMatchStats = 'playerMatchStats',
  MatchLineUps = 'matchLineUps',
  MatchLineUpsPlayers = 'matchLineUpsPlayers',
  FullMatchTeamsMatchStats = 'fullMatchTeamStats',
  FirstHalfTeamStats = 'firstHalfTeamStats',
  SecondHalfTeamStats = 'secondHalfTeamStats'
}

export enum ENTITIES_VIEWS {
  COUNTRIES = 'countries',
  STADIUMS = 'stadiums',
  TEAMS = 'teams',
  POSITIONS = 'positions',
  PLAYERS = 'players',
  REFEREES = 'referees',
  COMPETITIONS = 'competitions',
  SEASONS = 'seasons',
  MATCHES = 'matches',
  GOALS = 'goals'
}